improvement on CI and regression on features
removed feed generation and scssify as they crash now. Why?
maybe an auto update with breaking changes.
To investigate, but I wanted a blogging platform resistant to time, static means you don't have to update it!

Improved Dockerfile and compose to auto-update on development and added a script for generating a release build.

Tweaked exclude rule as apparently I had a python script at the root of my website :| .